Move site-specific settings out of settings.py.
Moved stuff that should be kept private or secret, or things that depend
on the deployment, out of settings.py and into a file that is not
tracked by version control. The README will indicate what needs to be in
that file.

NOTE: Anyone spelunking through the Git history (Hi there!) will note
that the commits before this point have the SECRET_KEY in them. When I
pulled it out, I also changed it. I'm not a moron.